+++ ncp/stores/__init__.py
@@ -1,2 +1,7 @@
 """Store backends package."""
 
+from ncp.stores.pgvector import PgvectorStore
+from ncp.stores.redis import RedisStore
+from ncp.stores.sqlite import SQLiteStore
+
+__all__ = ["SQLiteStore", "RedisStore", "PgvectorStore"]